updated not to show the progress indicator when there are no records in case of the module list views.
Updated not to show the progress indicator when there are no records in a module or while trying to filter the records in a module. Also updated to show the module name for which the records are not available in case there are no results. In module details screen, updated the code to append the value of the title fields to the title only after checking the value for null.