added tablet support by creating fragments for reuse by tab and phone.
added tablet support by creating fragments for reuse by tab and phone.
The following files have been renamed to keep the history intact for the
smart phone Activities used earlier.

1. renamed ModuleDetailsActivity to ModuleDetailFragment.
2. renamed EditModuleDetailsActivity to EditModuleDetailFragment.
3. renamed RecentListActivity to RecentModuleListFragment.

In all fragments, removed setContentView and create the view in
onCreateView and added fragment lifecycle methods updated the protected
signatures of Activity for resume and stop to public. The design borrows
heavily from IOSched app, but also keeps the core logic built-in earlier
as is by including onActivityCreated.

TODO: 
1. The header methods will fail in the new single pane and
multi-pane activities , so dummy stubs have been added as a stop gap
till the UI for tablets and phones is standard-ized. 
2.The menus and actionbar settings are still pending for the same reason
as above.