fix for the Invalid sessionId error from sugar crm server
fix for the Invalid sessionId error from sugar crm server. after prolonged periods if app is
opened, the session gets expired and is not handled by Rest API's. To circumvent that,
session id expiration is currently handled in SugarCrmApp.getSessionId based on a fixed
time for renewing the session. If this fixed time is not enough then the
 "Invalid session Ids" are not propagated back, so we keep the time low for
the time being and check for sessionId when updating the data on the server.